Create invisible.py
The cloak or the cloth color used gets replaced by the surroundings captured for 2sec. This makes the output look like the user is performing the harry potter cloak technique.

Note: Avoid colors like skin, red, black, or pink. Choose more rare or unnatural colors so that the user cannot get vanished.
Green screen color is rarely matched with surroundings, which makes it the best to mask people as in these programs.